feature+refactor: add dql and tag search
- move rest api proxy functions to new files
