+++ packages/mcp-server/src/features/local-rest-api/active-file.ts
@@ -1,0 +1,132 @@
+import { f, makeRequest, type ToolRegistry } from "$/shared";
+import type { Server } from "@modelcontextprotocol/sdk/server/index.js";
+import { type } from "arktype";
+import { LocalRestAPI } from "shared";
+
+export default function setup(tools: ToolRegistry, server: Server) {
+  // GET Active File
+  tools.register(
+    type({
+      name: '"get_active_file"',
+      arguments: {
+        format: type('"markdown" | "json"').optional(),
+      },
+    }).describe(
+      "Returns the content of the currently active file in Obsidian. Can return either markdown content or a JSON representation including parsed tags and frontmatter.",
+    ),
+    async ({ arguments: args }) => {
+      const format =
+        args?.format === "json"
+          ? "application/vnd.olrapi.note+json"
+          : "text/markdown";
+      const data = await makeRequest(
+        LocalRestAPI.ApiNoteJson.or("string"),
+        "/active/",
+        {
+          headers: { Accept: format },
+        },
+      );
+      const content =
+        typeof data === "string" ? data : JSON.stringify(data, null, 2);
+      return { content: [{ type: "text", text: content }] };
+    },
+  );
+
+  // PUT Active File
+  tools.register(
+    type({
+      name: '"update_active_file"',
+      arguments: {
+        content: "string",
+      },
+    }).describe("Update the content of the active file open in Obsidian."),
+    async ({ arguments: args }) => {
+      await makeRequest(LocalRestAPI.ApiNoContentResponse, "/active/", {
+        method: "PUT",
+        body: args.content,
+      });
+      return {
+        content: [{ type: "text", text: "File updated successfully" }],
+      };
+    },
+  );
+
+  // POST Active File
+  tools.register(
+    type({
+      name: '"append_to_active_file"',
+      arguments: {
+        content: "string",
+      },
+    }).describe("Append content to the end of the currently-open note."),
+    async ({ arguments: args }) => {
+      await makeRequest(LocalRestAPI.ApiNoContentResponse, "/active/", {
+        method: "POST",
+        body: args.content,
+      });
+      return {
+        content: [{ type: "text", text: "Content appended successfully" }],
+      };
+    },
+  );
+
+  // PATCH Active File
+  tools.register(
+    type({
+      name: '"patch_active_file"',
+      arguments: LocalRestAPI.ApiPatchParameters,
+    }).describe(
+      "Insert or modify content in the currently-open note relative to a heading, block reference, or frontmatter field.",
+    ),
+    async ({ arguments: args }) => {
+      const headers: Record<string, string> = {
+        Operation: args.operation,
+        "Target-Type": args.targetType,
+        Target: args.target,
+        "Create-Target-If-Missing": "true",
+      };
+
+      if (args.targetDelimiter) {
+        headers["Target-Delimiter"] = args.targetDelimiter;
+      }
+      if (args.trimTargetWhitespace !== undefined) {
+        headers["Trim-Target-Whitespace"] = String(args.trimTargetWhitespace);
+      }
+      if (args.contentType) {
+        headers["Content-Type"] = args.contentType;
+      }
+
+      const response = await makeRequest(
+        LocalRestAPI.ApiContentResponse,
+        "/active/",
+        {
+          method: "PATCH",
+          headers,
+          body: args.content,
+        },
+      );
+      return {
+        content: [
+          { type: "text", text: "File patched successfully" },
+          { type: "text", text: response },
+        ],
+      };
+    },
+  );
+
+  // DELETE Active File
+  tools.register(
+    type({
+      name: '"delete_active_file"',
+      arguments: "Record<string, unknown>",
+    }).describe("Delete the currently-active file in Obsidian."),
+    async () => {
+      await makeRequest(LocalRestAPI.ApiNoContentResponse, "/active/", {
+        method: "DELETE",
+      });
+      return {
+        content: [{ type: "text", text: "File deleted successfully" }],
+      };
+    },
+  );
+}
