+++ packages/mcp-server/src/features/local-rest-api/vault.ts
@@ -1,0 +1,180 @@
+import { makeRequest, type ToolRegistry } from "$/shared";
+import type { Server } from "@modelcontextprotocol/sdk/server/index.js";
+import { type } from "arktype";
+import { LocalRestAPI } from "shared";
+
+export default function setup(tools: ToolRegistry, server: Server) {
+  // GET Vault Files or Directories List
+  tools.register(
+    type({
+      name: '"list_vault_files"',
+      arguments: {
+        "directory?": "string",
+      },
+    }).describe(
+      "List files in the root directory or a specified subdirectory of your vault.",
+    ),
+    async ({ arguments: args }) => {
+      const path = args.directory ? `${args.directory}/` : "";
+      const data = await makeRequest(
+        LocalRestAPI.ApiVaultFileResponse.or(
+          LocalRestAPI.ApiVaultDirectoryResponse,
+        ),
+        `/vault/${path}`,
+      );
+      return {
+        content: [{ type: "text", text: JSON.stringify(data, null, 2) }],
+      };
+    },
+  );
+
+  // GET Vault File Content
+  tools.register(
+    type({
+      name: '"get_vault_file"',
+      arguments: {
+        filename: "string",
+        "format?": '"markdown" | "json"',
+      },
+    }).describe("Get the content of a file from your vault."),
+    async ({ arguments: args }) => {
+      const isJson = args.format === "json";
+      const format = isJson
+        ? "application/vnd.olrapi.note+json"
+        : "text/markdown";
+      const data = await makeRequest(
+        isJson ? LocalRestAPI.ApiNoteJson : LocalRestAPI.ApiContentResponse,
+        `/vault/${encodeURIComponent(args.filename)}`,
+        {
+          headers: { Accept: format },
+        },
+      );
+      return {
+        content: [
+          {
+            type: "text",
+            text:
+              typeof data === "string" ? data : JSON.stringify(data, null, 2),
+          },
+        ],
+      };
+    },
+  );
+
+  // PUT Vault File Content
+  tools.register(
+    type({
+      name: '"create_vault_file"',
+      arguments: {
+        filename: "string",
+        content: "string",
+      },
+    }).describe("Create a new file in your vault or update an existing one."),
+    async ({ arguments: args }) => {
+      await makeRequest(
+        LocalRestAPI.ApiNoContentResponse,
+        `/vault/${encodeURIComponent(args.filename)}`,
+        {
+          method: "PUT",
+          body: args.content,
+        },
+      );
+      return {
+        content: [{ type: "text", text: "File created successfully" }],
+      };
+    },
+  );
+
+  // POST Vault File Content
+  tools.register(
+    type({
+      name: '"append_to_vault_file"',
+      arguments: {
+        filename: "string",
+        content: "string",
+      },
+    }).describe("Append content to a new or existing file."),
+    async ({ arguments: args }) => {
+      await makeRequest(
+        LocalRestAPI.ApiNoContentResponse,
+        `/vault/${encodeURIComponent(args.filename)}`,
+        {
+          method: "POST",
+          body: args.content,
+        },
+      );
+      return {
+        content: [{ type: "text", text: "Content appended successfully" }],
+      };
+    },
+  );
+
+  // PATCH Vault File Content
+  tools.register(
+    type({
+      name: '"patch_vault_file"',
+      arguments: type({
+        filename: "string",
+      }).and(LocalRestAPI.ApiPatchParameters),
+    }).describe(
+      "Insert or modify content in a file relative to a heading, block reference, or frontmatter field.",
+    ),
+    async ({ arguments: args }) => {
+      const headers: HeadersInit = {
+        Operation: args.operation,
+        "Target-Type": args.targetType,
+        Target: args.target,
+        "Create-Target-If-Missing": "true",
+      };
+
+      if (args.targetDelimiter) {
+        headers["Target-Delimiter"] = args.targetDelimiter;
+      }
+      if (args.trimTargetWhitespace !== undefined) {
+        headers["Trim-Target-Whitespace"] = String(args.trimTargetWhitespace);
+      }
+      if (args.contentType) {
+        headers["Content-Type"] = args.contentType;
+      }
+
+      const response = await makeRequest(
+        LocalRestAPI.ApiContentResponse,
+        `/vault/${encodeURIComponent(args.filename)}`,
+        {
+          method: "PATCH",
+          headers,
+          body: args.content,
+        },
+      );
+
+      return {
+        content: [
+          { type: "text", text: "File patched successfully" },
+          { type: "text", text: response },
+        ],
+      };
+    },
+  );
+
+  // DELETE Vault File Content
+  tools.register(
+    type({
+      name: '"delete_vault_file"',
+      arguments: {
+        filename: "string",
+      },
+    }).describe("Delete a file from your vault."),
+    async ({ arguments: args }) => {
+      await makeRequest(
+        LocalRestAPI.ApiNoContentResponse,
+        `/vault/${encodeURIComponent(args.filename)}`,
+        {
+          method: "DELETE",
+        },
+      );
+      return {
+        content: [{ type: "text", text: "File deleted successfully" }],
+      };
+    },
+  );
+}
